feat(sql): Refactor SQL command structure by introducing BaseDatabaseCommand and BaseServerCommand; update index recommendation logic and options handling
