+++ src/Commands/Sql/SqlJsonContext.cs
@@ -3,12 +3,12 @@
 
 using System.Text.Json.Serialization;
 using AzureMcp.Commands.Sql.Database;
-using AzureMcp.Commands.Sql.Index;
+using AzureMcp.Commands.Sql.Database.Index;
 using AzureMcp.Commands.Sql.Server;
 
 namespace AzureMcp.Commands.Sql;
 
-[JsonSerializable(typeof(Index.SqlIndexRecommendCommand.IndexRecommendCommandResult))]
+[JsonSerializable(typeof(SqlIndexRecommendCommand.IndexRecommendCommandResult))]
 [JsonSerializable(typeof(SqlDatabaseListCommand.DatabaseListCommandResult))]
 [JsonSerializable(typeof(SqlServerListCommand.ServerListCommandResult))]
 [JsonSerializable(typeof(Database.SqlDatabaseListCommand.DatabaseListCommandResult))]
