+++ src/Models/Sql/IIndexRecommendCommandResult.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Collections.Generic;
-
 namespace AzureMcp.Models.Sql;
 
 internal interface IIndexRecommendCommandResult
